Updated default enum in manure module
